LayoutTests/imported/w3c/web-platform-tests/dom/events/ProgressEvent-expected.txt


FAIL Default event values. assert_equals: expected (boolean) false but got (undefined) undefined
PASS document.createEvent() should not work with ProgressEvent. 
